refactor: move product-variant to node-app
