+++ imports/node-app/plugins/product-variant/i18n/de.json
