+++ imports/node-app/plugins/product-variant/i18n/pl.json
