+++ imports/node-app/plugins/product-variant/i18n/zh.json
